PROGRES : Methodes commentaires

+++ Serveur/Services/CommentaireService.cs
@@ -16,8 +16,16 @@ namespace arsoudeServeur.Services
 
         public void DeleteCommentaire(int id, Utilisateur utilisateurCourant)
         {
-           _context.commentaires.Remove(_context.commentaires.Find(id));
-            _context.SaveChangesAsync();
+            var commentaire = _context.commentaires.Where(c => c.id == id).FirstOrDefault();
+            if (utilisateurCourant.id == commentaire.utilisateurId || utilisateurCourant.role == "Administrator")
+            {
+                _context.commentaires.Remove(_context.commentaires.Find(id));
+                _context.SaveChangesAsync();
+            }
+            else
+            {
+                //Vous n'êtes pas l'auteur du commentaire ou vous n'êtes pas un admin
+            }
         }
 
         public IEnumerable<Commentaire> GetCommentaires(int id)
@@ -34,34 +42,39 @@ namespace arsoudeServeur.Services
             _context.commentaires.Add(new Commentaire()
             {
                 texte = commentaire.texte,
-                utilisateur = utilisateurCourant,
                 review = commentaire.review,
                 randonnee = rando,
                 randonneeId = rando.id,
+                utilisateur = utilisateurCourant,
                 utilisateurId = utilisateurCourant.id
-
-                
             });
 
             _context.SaveChangesAsync();
         }
 
-        public void PutCommentaire(int id, CommentaireDTO commentaire, Utilisateur utilisateurCourant)
+        public void PutCommentaire(int id, CommentaireDTO commentaireDTO, Utilisateur utilisateurCourant)
         {
-            var rando = _context.randonnees.Where(r => r.id == commentaire.randoId).FirstOrDefault();    
-
-            _context.commentaires.Update(new Commentaire()
+            var rando = _context.randonnees.Where(r => r.id == commentaireDTO.randoId).FirstOrDefault();
+            var commentaire = _context.commentaires.Where(c => c.id == id).FirstOrDefault();
+            if (utilisateurCourant.id == commentaire.utilisateurId || utilisateurCourant.role == "Administrator")
             {
-                id = id,
-                texte = commentaire.texte,
-                utilisateur = utilisateurCourant,
-                review = commentaire.review,
-                randonnee = rando,
-                randonneeId = rando.id,
-                utilisateurId = utilisateurCourant.id
-            });
+                _context.commentaires.Update(new Commentaire()
+                {
+                    id = id,
+                    texte = commentaire.texte,
+                    utilisateur = utilisateurCourant,
+                    review = commentaire.review,
+                    randonnee = rando,
+                    randonneeId = rando.id,
+                    utilisateurId = utilisateurCourant.id
+                });
 
-            _context.SaveChangesAsync();
+                _context.SaveChangesAsync();
+            }
+            else
+            {
+                //Unauthorized ?
+            }
 
         }
     }
